Extract blade tests verbatim
