+++ tests/Blade/BladeEndSectionTest.php
@@ -1,0 +1,24 @@
+<?php
+
+use Mockery as m;
+use PHPUnit\Framework\TestCase;
+use Illuminate\View\Compilers\BladeCompiler;
+
+class BladeStopTest extends TestCase
+{
+    public function tearDown()
+    {
+        m::close();
+    }
+
+    public function testStopSectionsAreCompiled()
+    {
+        $compiler = new BladeCompiler($this->getFiles(), __DIR__);
+        $this->assertEquals('<?php $__env->stopSection(); ?>', $compiler->compileString('@stop'));
+    }
+
+    protected function getFiles()
+    {
+        return m::mock('Illuminate\Filesystem\Filesystem');
+    }
+}
