+++ tests/Blade/BladePushTest.php
@@ -1,0 +1,30 @@
+<?php
+
+use Mockery as m;
+use PHPUnit\Framework\TestCase;
+use Illuminate\View\Compilers\BladeCompiler;
+
+class BladePushTest extends TestCase
+{
+    public function tearDown()
+    {
+        m::close();
+    }
+
+    public function testPushIsCompiled()
+    {
+        $compiler = new BladeCompiler($this->getFiles(), __DIR__);
+        $string = '@push(\'foo\')
+test
+@endpush';
+        $expected = '<?php $__env->startPush(\'foo\'); ?>
+test
+<?php $__env->stopPush(); ?>';
+        $this->assertEquals($expected, $compiler->compileString($string));
+    }
+
+    protected function getFiles()
+    {
+        return m::mock('Illuminate\Filesystem\Filesystem');
+    }
+}
